requirements.txt
aiohttp>=3.9
aiosqlite>=0.19
fastapi>=0.110
python-dotenv>=1.0
uvicorn[standard]>=0.29
web3>=6.15
py-clob-client>=0.17
websockets>=12.0
